Create README.md
Any Thought or Idea, Knowledgeable Advice, Expertise Contributes, Professional Service or Supportive Efforts are Welcome and Appreciated. Your Knowledge, Talents, and Experience is Encouraged to Contribute or Join in, an effort to broaden opportunities that showcase creative talent, allow exposure and growth for independent artists to thrive, and continue to display an artistic experience for our entertainment. Let the human mind wonder toward enjoyment, Saturate our dreamscapes with beauty, surround communities with a kinetic energy that connects us and makes life on earth so magical.